Add request/response validation
Use the `openapi3filter` functions for request and response validation
to make sure that the traffic we're sending and receiving from the
server in the test is according to spec.

This code needs some cleanup and rework still, but the basic case is
back to working. Especially the error handling/recording here needs a
second look to give better ideas on what request/response was
problematic.