Fix request body validation
The request body was validated only when it had already been read, since
the request had already been sent. As kin-openapi is expected to be used
on the server side, the GetBody function was not used as it's expected
to be nil. Allowing for proper body validation by setting the body
explicitly back to the original by using `GetBody` when recording the
request/response.

Add more tests for the base cases.