Wrapped errors, and skipping 500s
Set the verifier to skip 500 responses by default, and add an option to
start the client with this verification as well if that should be needed
for anything.

Errors are now wrapped into a VerificationError struct, that allows a
sentinel type to be included for easy `errors.Is` checking of the type
of error that happened.