Allow skipping of request body validation
The body validation of requests conflicts with provoking specifically
bad request errors, so disabling it by default to allow the test client
to cause those type of errors as well. Should the validation be
required, the option "WithRequestBodyValidation" should be used.

Breaking change: If using the NewVerifier separately, the constructor
has now been changed to take the same functional arguments as the
client. This means that any usage of `basePath` must be updated to use
the Option `WithBasePath` instead.