Port copper to libopenapi
It very much seems like kin-openapi support for OpenAPI v3.1 will never
actually be done, so switching the implementation to use libopenapi
instead.

This requires fairly large rewrite of the matching logic, but the API of
copper has remained largely the same, with the exception of a couple of
functional options.

The errors are now also more verbose, and more descriptive since
libopenapi adds more context and even suggests changes to the spec.